requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
# Editable install with no version control (fast-api-builder==0.1.0)
-e ../
fastapi==0.115.0
graphql-core==3.2.4
h11==0.14.0
idna==3.10
iso8601==1.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
more-itertools==10.5.0
pluralize==20240519.3
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.10
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
strawberry-graphql==0.243.0
tomlkit==0.13.2
tortoise-orm==0.21.6
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wheel==0.44.0
